@@ -6,7 +6,7 @@ repository: https://github.com/Flutterista/platform_widgets
 issue_tracker: https://github.com/Flutterista/platform_widgets/issues
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
