@@ -14,7 +14,6 @@ class PlatformBar extends PlatformWidget<CupertinoNavigationBar, AppBar> impleme
     this.cupertinoPreviousPageTitle,
     this.cupertinoPadding,
     this.cupertinoBorder,
-    this.cupertinoActionsForegroundColor,
     this.cupertinoTransitionBetweenRoutes = true,
     this.materialActions,
     this.materialFlexibleSpace,
@@ -31,31 +30,30 @@ class PlatformBar extends PlatformWidget<CupertinoNavigationBar, AppBar> impleme
     this.materialBottomOpacity = 1.0,
   });
 
-  final Widget middle;
-  final Widget leading;
+  final Widget? middle;
+  final Widget? leading;
   final bool automaticallyImplyLeading;
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   /// Only used in NavigationBar
-  final Widget cupertinoTrailing;
+  final Widget? cupertinoTrailing;
   final bool cupertinoAutomaticallyImplyMiddle;
-  final String cupertinoPreviousPageTitle;
-  final EdgeInsetsDirectional cupertinoPadding;
-  final Border cupertinoBorder;
-  final Color cupertinoActionsForegroundColor;
+  final String? cupertinoPreviousPageTitle;
+  final EdgeInsetsDirectional? cupertinoPadding;
+  final Border? cupertinoBorder;
   final bool cupertinoTransitionBetweenRoutes;
 
   /// Only used in AppBar
-  final List<Widget> materialActions;
-  final Widget materialFlexibleSpace;
-  final PreferredSizeWidget materialBottom;
-  final double materialElevation;
-  final ShapeBorder materialShape;
-  final IconThemeData materialIconTheme;
-  final IconThemeData materialActionsIconTheme;
-  final TextTheme materialTextTheme;
+  final List<Widget>? materialActions;
+  final Widget? materialFlexibleSpace;
+  final PreferredSizeWidget? materialBottom;
+  final double? materialElevation;
+  final ShapeBorder? materialShape;
+  final IconThemeData? materialIconTheme;
+  final IconThemeData? materialActionsIconTheme;
+  final TextTheme? materialTextTheme;
   final bool materialPrimary;
-  final bool materialCenterTitle;
+  final bool? materialCenterTitle;
   final double materialTitleSpacing;
   final double materialToolbarOpacity;
   final double materialBottomOpacity;
@@ -71,7 +69,6 @@ class PlatformBar extends PlatformWidget<CupertinoNavigationBar, AppBar> impleme
       automaticallyImplyMiddle: cupertinoAutomaticallyImplyMiddle,
       padding: cupertinoPadding,
       border: cupertinoBorder,
-      actionsForegroundColor: cupertinoActionsForegroundColor,
       transitionBetweenRoutes: cupertinoTransitionBetweenRoutes,
     );
   }
